Fix some inconsistency
